Remove versions from packages
