+++ bundles/pixi.js-legacy/package.json
@@ -1,6 +1,5 @@
 {
   "name": "pixi.js-legacy",
-  "version": "7.1.2",
   "description": "The HTML5 Creation Engine: Create beautiful digital content with the fastest, most flexible 2D WebGL renderer.",
   "author": "Mat Groves",
   "contributors": [
@@ -47,16 +46,16 @@
     "url": "https://opencollective.com/pixijs"
   },
   "dependencies": {
-    "@pixi/canvas-display": "7.1.2",
-    "@pixi/canvas-extract": "7.1.2",
-    "@pixi/canvas-graphics": "7.1.2",
-    "@pixi/canvas-mesh": "7.1.2",
-    "@pixi/canvas-particle-container": "7.1.2",
-    "@pixi/canvas-prepare": "7.1.2",
-    "@pixi/canvas-renderer": "7.1.2",
-    "@pixi/canvas-sprite": "7.1.2",
-    "@pixi/canvas-sprite-tiling": "7.1.2",
-    "@pixi/canvas-text": "7.1.2",
-    "pixi.js": "7.1.2"
+    "@pixi/canvas-display": "file:../../packages/canvas-display",
+    "@pixi/canvas-extract": "file:../../packages/canvas-extract",
+    "@pixi/canvas-graphics": "file:../../packages/canvas-graphics",
+    "@pixi/canvas-mesh": "file:../../packages/canvas-mesh",
+    "@pixi/canvas-particle-container": "file:../../packages/canvas-particle-container",
+    "@pixi/canvas-prepare": "file:../../packages/canvas-prepare",
+    "@pixi/canvas-renderer": "file:../../packages/canvas-renderer",
+    "@pixi/canvas-sprite": "file:../../packages/canvas-sprite",
+    "@pixi/canvas-sprite-tiling": "file:../../packages/canvas-sprite-tiling",
+    "@pixi/canvas-text": "file:../../packages/canvas-text",
+    "pixi.js": "file:../pixi.js"
   }
 }
