+++ bundles/pixi.js-node/package.json
@@ -1,6 +1,5 @@
 {
   "name": "@pixi/node",
-  "version": "7.1.2",
   "description": "Bundle for PixiJS with support for NodeJS",
   "homepage": "http://pixijs.com/",
   "bugs": "https://github.com/pixijs/pixijs/issues",
@@ -40,32 +39,32 @@
     "coverage": "npm run test -- --coverage --maxWorkers=4"
   },
   "dependencies": {
-    "@pixi/app": "7.1.2",
-    "@pixi/assets": "7.1.2",
-    "@pixi/core": "7.1.2",
-    "@pixi/display": "7.1.2",
-    "@pixi/extensions": "7.1.2",
-    "@pixi/extract": "7.1.2",
-    "@pixi/filter-alpha": "7.1.2",
-    "@pixi/filter-blur": "7.1.2",
-    "@pixi/filter-color-matrix": "7.1.2",
-    "@pixi/filter-displacement": "7.1.2",
-    "@pixi/filter-fxaa": "7.1.2",
-    "@pixi/filter-noise": "7.1.2",
-    "@pixi/graphics": "7.1.2",
-    "@pixi/mesh": "7.1.2",
-    "@pixi/mesh-extras": "7.1.2",
-    "@pixi/mixin-cache-as-bitmap": "7.1.2",
-    "@pixi/mixin-get-child-by-name": "7.1.2",
-    "@pixi/mixin-get-global-position": "7.1.2",
-    "@pixi/particle-container": "7.1.2",
-    "@pixi/prepare": "7.1.2",
-    "@pixi/sprite": "7.1.2",
-    "@pixi/sprite-animated": "7.1.2",
-    "@pixi/sprite-tiling": "7.1.2",
-    "@pixi/spritesheet": "7.1.2",
-    "@pixi/text": "7.1.2",
-    "@pixi/text-bitmap": "7.1.2",
+    "@pixi/app": "file:../../packages/app",
+    "@pixi/assets": "file:../../packages/assets",
+    "@pixi/core": "file:../../packages/core",
+    "@pixi/display": "file:../../packages/display",
+    "@pixi/extensions": "file:../../packages/extensions",
+    "@pixi/extract": "file:../../packages/extract",
+    "@pixi/filter-alpha": "file:../../packages/filter-alpha",
+    "@pixi/filter-blur": "file:../../packages/filter-blur",
+    "@pixi/filter-color-matrix": "file:../../packages/filter-color-matrix",
+    "@pixi/filter-displacement": "file:../../packages/filter-displacement",
+    "@pixi/filter-fxaa": "file:../../packages/filter-fxaa",
+    "@pixi/filter-noise": "file:../../packages/filter-noise",
+    "@pixi/graphics": "file:../../packages/graphics",
+    "@pixi/mesh": "file:../../packages/mesh",
+    "@pixi/mesh-extras": "file:../../packages/mesh-extras",
+    "@pixi/mixin-cache-as-bitmap": "file:../../packages/mixin-cache-as-bitmap",
+    "@pixi/mixin-get-child-by-name": "file:../../packages/mixin-get-child-by-name",
+    "@pixi/mixin-get-global-position": "file:../../packages/mixin-get-global-position",
+    "@pixi/particle-container": "file:../../packages/particle-container",
+    "@pixi/prepare": "file:../../packages/prepare",
+    "@pixi/sprite": "file:../../packages/sprite",
+    "@pixi/sprite-animated": "file:../../packages/sprite-animated",
+    "@pixi/sprite-tiling": "file:../../packages/sprite-tiling",
+    "@pixi/spritesheet": "file:../../packages/spritesheet",
+    "@pixi/text": "file:../../packages/text",
+    "@pixi/text-bitmap": "file:../../packages/text-bitmap",
     "@types/gl": "^6.0.2",
     "@xmldom/xmldom": "^0.8.6",
     "canvas": "^2.9.1",
