+++ package.json
@@ -38,6 +38,7 @@
     "postdist": "copyfiles -f bundles/*/dist/* dist && copyfiles -f packages/*/dist/* dist/packages",
     "prerelease": "run-s clean:build test dist",
     "release": "ts-node ./scripts/release.ts",
+    "prepublish-ci": "ts-node ./scripts/prepublish.ts",
     "publish-ci": "workspaces-run --only-fs=\"{packages,bundles}/*\" -- npm publish",
     "webdoc": "webdoc"
   },
