+++ packages/basis/package.json
@@ -1,6 +1,5 @@
 {
   "name": "@pixi/basis",
-  "version": "7.1.2",
   "description": "Loader for the *.basis supercompressed texture file format. This package also ships with the transcoder!",
   "keywords": [
     "basisu",
