+++ packages/compressed-textures/package.json
@@ -1,6 +1,5 @@
 {
   "name": "@pixi/compressed-textures",
-  "version": "7.1.2",
   "description": "Loaders for compressed texture file formats",
   "keywords": [
     "pixi.js",
