+++ packages/math/package.json
@@ -1,6 +1,5 @@
 {
   "name": "@pixi/math",
-  "version": "7.1.2",
   "main": "lib/index.js",
   "module": "lib/index.mjs",
   "types": "lib/index.d.ts",
