+++ packages/settings/package.json
@@ -1,6 +1,5 @@
 {
   "name": "@pixi/settings",
-  "version": "7.1.2",
   "main": "lib/index.js",
   "module": "lib/index.mjs",
   "types": "lib/index.d.ts",
@@ -36,7 +35,7 @@
     "*.d.ts"
   ],
   "dependencies": {
-    "@pixi/constants": "7.1.2",
+    "@pixi/constants": "file:../constants",
     "@types/css-font-loading-module": "^0.0.7"
   },
   "devDependencies": {
