+++ packages/utils/package.json
@@ -1,6 +1,5 @@
 {
   "name": "@pixi/utils",
-  "version": "7.1.2",
   "main": "lib/index.js",
   "module": "lib/index.mjs",
   "types": "lib/index.d.ts",
@@ -36,9 +35,9 @@
     "*.d.ts"
   ],
   "dependencies": {
-    "@pixi/color": "7.1.2",
-    "@pixi/constants": "7.1.2",
-    "@pixi/settings": "7.1.2",
+    "@pixi/color": "file:../color",
+    "@pixi/constants": "file:../constants",
+    "@pixi/settings": "file:../settings",
     "@types/earcut": "^2.1.0",
     "earcut": "^2.2.4",
     "eventemitter3": "^4.0.0",
